+++ src/IR/irGen/expressions/ContainerAssignmentExpressionGenerationStrategy/ContainerAssignmentExpressionGenerationStrategy.h
@@ -23,14 +23,14 @@ class ContainerAssignmentExpressionGenerationStrategy
                                 llvm::Value *&rhsVariable,
                                 llvm::ArrayType *&rhsArrayType,
                                 llvm::Type *arrayElementType,
-                                const std::vector<size_t> &lhsSizes,
-                                const std::vector<size_t> &rhsSizes);
+                                const std::vector<uint64_t> &lhsSizes,
+                                const std::vector<uint64_t> &rhsSizes);
 
   void assignArray(llvm::ArrayType *&arrayType, llvm::Value *&variable,
                    llvm::Value *&rhsVariable, llvm::ArrayType *&rhsArrayType,
                    llvm::Type *rhsArrayElementType,
                    std::vector<llvm::Value *> &indices,
-                   const std::vector<size_t> &rhsSizes, uint64_t index);
+                   const std::vector<uint64_t> &rhsSizes, uint64_t index);
 
  private:
   // LHS
@@ -44,7 +44,7 @@ class ContainerAssignmentExpressionGenerationStrategy
   llvm::ArrayType *_rhsArrayType;
   llvm::Value *_rhsSize;
 
-  std::vector<size_t> _lhsSizes, _rhsSizes;
+  std::vector<uint64_t> _lhsSizes, _rhsSizes;
   llvm::Type *_rhsArrayElementType, *_lhsArrayElementType;
 };
